requirements.txt

orjson
dash_treeview_antd
dash==2.11.1
dash-bootstrap-components==1.4.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.0
dash-table==5.0.0
dash-daq==0.5.0

flask-cors==3.0.10
Flask-Admin==1.6.0
flask-sqlalchemy==2.5.1
Flask-Migrate==3.1.0
sqlalchemy-mixins==1.5.3
sqlalchemy==1.4.45
slugify

pandasdmx
pandasdmx[cache]
scipy
openpyxl
shapely

sentry-sdk[flask]

flask-login
requests-cache
